+++ PPI/ppi_2_2.R
@@ -1,0 +1,9 @@
+A=matrix(1:6,2,3)
+B=matrix(c(1,2,1,2,1,2),2,3)
+C=matrix(c(4,3,-1,-2,7,1),2,3)
+a=((3*A)+B)%*%t(C-(2*B))
+a
+b=(A+B)*(A+B)
+b
+c=det(t(A)%*%A)
+c
