+++ PPI/ppi_2_3.R
@@ -1,0 +1,6 @@
+A=matrix(c(3,5,-2,-4),2,2)
+B=matrix(c(-1,-5,2,6),2,2)
+X=B/A
+A
+B
+X
