+++ PPI/ppi_2_5.R
@@ -1,0 +1,2 @@
+samochody=data.frame(marka=c("Honda","Honda","Mazda","Opel"),rocznik=c(2003,2005,2007,2001),uszkodzony=c("TRUE","FALSE","FALSE","TRUE"),kolor=c("czerwony","srebrny","zielony","niebieski"))
+samochody
