+++ PPI/ppi_3_1.R
@@ -1,0 +1,2 @@
+wektor=c("A.1","B.2","C.3","D.4","A.5","B.6","C.7","D.8","A.9","B.10","C.11","D.12","A.13","B.14","C.15","D.16","A.17","B.18","C.19","D.20")
+wektor
