+++ PPI/ppi_3_2.R
@@ -1,0 +1,18 @@
+wektor=c(1,2,3,4,5,6,7,8,9,10,NA,12)
+obj=data.frame(liczba=wektor,kwadrat=c(1,4,9,16,25,36,49,64,81,100,NA,144))
+obj
+class(obj)
+mode(obj)
+typeof(obj)
+str(obj)
+object.size(obj)
+is.na(obj)
+fix(obj)
+obj
+fix(obj)
+obj
+wektor3<-c(1,8,27,64,125,216,343,512,729,1000,1331,1728)
+object<-data.frame(x=obj[1],x^2=obj[2],x^3=wektor3)
+object
+is.matrix(object)
+class(object)
