+++ Stata/stata_conf.R
@@ -1,0 +1,7 @@
+#Bauer 99029
+
+install.packages("dplyr")
+install.packages("devtools")
+devtools::install_github("pbiecek/PrzewodnikPakiet")
+library(dplyr)
+library(PogromcyDanych)
